+++ src/seamlesserverpkg/renderer/base/karaxNodes/dynamicLink.nim
@@ -4,11 +4,12 @@ when defined js:
 
 include pkg/karax/prelude
 
-proc dynamicLink*(href = ""; title = ""): VNode =
+proc dynamicLink*(href = ""; title = ""; inSpa = true): VNode =
   ## Creates a default HTML link in backend and in frontend it just pushes the
-  ## state to history
+  ## state to history, if the link is in SPA
   buildHtml(a(href = href, title = title)):
     when defined js:
-      proc onclick(ev: Event; n: VNode) =
-        preventDefault ev
-        redirect href
+      if inSpa:
+        proc onclick(ev: Event; n: VNode) =
+          preventDefault ev
+          redirect href